+++ src/Federation/EntityResolverProvider.php
@@ -120,12 +120,7 @@ class EntityResolverProvider
             throw new Error(self::unknownTypename($typename));
         }
 
-        /**
-         * TODO remove when upgrading graphql-php.
-         *
-         * @var (\GraphQL\Language\AST\Node&\GraphQL\Language\AST\TypeDefinitionNode)|null $definition
-         */
-        $definition = $type->astNode;
+        $definition = $type->astNode();
         if (null === $definition) {
             throw new FederationException("Must provide AST definition for type `{$typename}`.");
         }