+++ src/Testing/TestingServiceProvider.php
@@ -4,6 +4,7 @@ namespace Nuwave\Lighthouse\Testing;
 
 use Illuminate\Contracts\Events\Dispatcher;
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Testing\TestResponse;
 use Nuwave\Lighthouse\Events\RegisterDirectiveNamespaces;
 
 class TestingServiceProvider extends ServiceProvider
@@ -11,12 +12,7 @@ class TestingServiceProvider extends ServiceProvider
     public function register(): void
     {
         $this->app->singleton(MockResolverService::class);
-
-        if (class_exists('Illuminate\Testing\TestResponse')) {
-            \Illuminate\Testing\TestResponse::mixin(new TestResponseMixin());
-        } elseif (class_exists('Illuminate\Foundation\Testing\TestResponse')) {
-            \Illuminate\Foundation\Testing\TestResponse::mixin(new TestResponseMixin());
-        }
+        TestResponse::mixin(new TestResponseMixin());
     }
 
     public function boot(Dispatcher $dispatcher): void