+++ src/Validation/ValidatorDirective.php
@@ -85,9 +85,7 @@ GRAPHQL;
     public function manipulateTypeDefinition(DocumentAST &$documentAST, TypeDefinitionNode &$typeDefinition)
     {
         if (! $typeDefinition instanceof InputObjectTypeDefinitionNode) {
-            throw new DefinitionException(
-                "Can not use @validator on non input type {$typeDefinition->name->value}."
-            );
+            throw new DefinitionException("Can not use @validator on non input type {$typeDefinition->getName()->value}.");
         }
 
         $this->setFullClassnameOnDirective(