+++ tests/Integration/Schema/Directives/ThrottleDirectiveTest.php
@@ -8,7 +8,6 @@ use Illuminate\Http\Response;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Exceptions\DirectiveException;
 use Nuwave\Lighthouse\Exceptions\RateLimitException;
-use Nuwave\Lighthouse\Support\AppVersion;
 use Tests\TestCase;
 use Tests\Utils\Queries\Foo;
 
@@ -32,10 +31,6 @@ final class ThrottleDirectiveTest extends TestCase
 
     public function testNamedLimiterReturnsRequest(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */ '
         type Query {
             foo: Int @throttle(name: "test")
@@ -62,10 +57,6 @@ final class ThrottleDirectiveTest extends TestCase
 
     public function testNamedLimiter(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */ '
         type Query {
             foo: Int @throttle(name: "test")