+++ tests/Unit/Auth/GuardDirectiveTest.php
@@ -130,6 +130,6 @@ final class GuardDirectiveTest extends TestCase
                 }
             }
             ')
-            ->assertGraphQLErrorCategory(AuthenticationException::CATEGORY);
+            ->assertGraphQLErrorMessage(AuthenticationException::MESSAGE);
     }
 }