+++ tests/Unit/Schema/Directives/ThrottleDirectiveTest.php
@@ -4,7 +4,6 @@ namespace Tests\Unit\Schema\Directives;
 
 use Illuminate\Cache\RateLimiter;
 use Illuminate\Cache\RateLimiting\Limit;
-use Nuwave\Lighthouse\Support\AppVersion;
 use Tests\TestCase;
 use Tests\Utils\Queries\Foo;
 
@@ -12,10 +11,6 @@ final class ThrottleDirectiveTest extends TestCase
 {
     public function testNamedLimiter(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */ '
         type Query {
             foo: Int @throttle(name: "test")
@@ -71,10 +66,6 @@ final class ThrottleDirectiveTest extends TestCase
 
     public function testUnlimitedNamedLimiter(): void
     {
-        if (AppVersion::below(8.0)) {
-            $this->markTestSkipped('Version less than 8.0 does not support named requests.');
-        }
-
         $this->schema = /** @lang GraphQL */ '
         type Query {
             foo: Int @throttle(name: "test")
@@ -82,12 +73,10 @@ final class ThrottleDirectiveTest extends TestCase
         ';
 
         $rateLimiter = $this->createMock(RateLimiter::class);
-        // @phpstan-ignore-next-line phpstan ignores markTestSkipped
         $rateLimiter->expects(self::atLeast(1))
             ->method('limiter')
             ->with('test')
             ->willReturn(static function () {
-                // @phpstan-ignore-next-line phpstan ignores markTestSkipped
                 return Limit::none();
             });